BUG: It should be cast to double.

+++ starviewer/src/core/decaycorrectionfactorformula.cpp
@@ -14,7 +14,7 @@ DecayCorrectionFactorFormula::~DecayCorrectionFactorFormula()
 
 double DecayCorrectionFactorFormula::compute(int radionuclideHalfLifeInSeconds, int timeLapseInSeconds)
 {
-    return qPow(2.0, -timeLapseInSeconds / radionuclideHalfLifeInSeconds);
+    return qPow(2.0, -timeLapseInSeconds / (double)radionuclideHalfLifeInSeconds);
 }
 
 } // End namespace udg
